infrastructure/testharness.js/worker-dedicated.sub.html

<!DOCTYPE HTML>
<html>
<head>
<meta name="variant" content="">
<meta name="variant" content="?remove-promise">
<title>Dedicated Worker Tests</title>
<script src="resources/variant-promise.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<h1>Dedicated Web Worker Tests</h1>
<p>Demonstrates running <tt>testharness</tt> based tests inside a dedicated web worker.
<p>The test harness is expected to fail due to an uncaught exception in one worker.</p>
<div id="log"></div>

<script>
test(function(t) {
        assert_true("Worker" in self, "Browser should support Workers");
    },
    "Browser supports Workers");

fetch_tests_from_worker(new Worker("resources/worker.js"));

fetch_tests_from_worker(new Worker("resources/worker-error.js"));

test(function(t) {
        assert_false(false, "False is false");
    },
    "Test running on main document.");
</script>
</body>
